Adjust PSP DRA splits for next big code share
