+++ CMakeLists.txt
@@ -193,7 +193,7 @@ set(SOURCE_FILES_DRA
     src/dra/63ED4.c
     src/dra/d_CD80.c
     src/dra/692E8.c
-    src/dra/6D59C.c
+    src/dra/6D584.c
     src/dra/6E42C.c
     src/dra/6FDF8.c
     src/dra/704D0.c
