+++ config/splat.hd.dra.yaml
@@ -97,7 +97,7 @@ segments:
       - [0xCA20, .data, cd]
       - [0xCE48, .data, d_CD80]
       - [0xCEB8, .data, 692E8]
-      - [0xCFDC, .data, 6D59C]
+      - [0xCFDC, .data, 6D584]
       - [0xD010, .data, 6E42C]
       - [0xD030, .data, 6FDF8]
       - [0xD038, .data, d_CF74]
@@ -190,7 +190,7 @@ segments:
       - [0x63D1C, c, 63ED4]
       - [0x67200, c, cd]
       - [0x69100, c, 692E8]
-      - [0x6D2E0, c, 6D59C]
+      - [0x6D2C8, c, 6D584]
       - [0x6E170, c, 6E42C]
       - [0x6FB3C, c, 6FDF8]
       - [0x7021C, c, 704D0]
