+++ config/splat.us.dra.yaml
@@ -101,7 +101,7 @@ segments:
       - [0xC958, .data, cd]
       - [0xCD80, .data, d_CD80]
       - [0xCDF8, .data, 692E8]
-      - [0xCF18, .data, 6D59C]
+      - [0xCF18, .data, 6D584]
       - [0xCF4C, .data, 6E42C]
       - [0xCF6C, .data, 6FDF8]
       - [0xCF74, .data, d_CF74]
@@ -256,7 +256,7 @@ segments:
       - [0x63ED4, c, 63ED4]
       - [0x673C0, c, cd]
       - [0x692E8, c, 692E8]
-      - [0x6D59C, c, 6D59C]
+      - [0x6D584, c, 6D584]
       - [0x6E42C, c, 6E42C]
       - [0x6FDF8, c, 6FDF8]
       - [0x704D0, c, 704D0]
