--- src/dra_psp/3A770.c
@@ -1,14 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3A770", UpdateAnim);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3A770", func_psp_09117538);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3A770", func_8010DFF0);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3A770", func_8010E0A8);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3A770", func_psp_091176D8);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3A770", func_psp_091176F0);
