--- src/dra_psp/3B240.c
@@ -1,24 +1,0 @@
-// SPDX-License-Identifier: AGPL-3.0-or-later
-#include "../dra/dra.h"
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_8010E3E0);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_09117C00);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_09118330);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_09118408);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_091187D8);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_091190E0);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_09119950);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_09119D00);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_0911A0B8);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_0911ABF8);
-
-INCLUDE_ASM("dra_psp/psp/dra_psp/3B240", func_psp_0911B630);
